# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "post": {
        "operationId": "ReportsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a report (Generic)",
        "tags": [
          "Reports"
        ]
      }
    },
    "/experiences/{id}/report": {
      "post": {
        "operationId": "ReportsController_reportExperience",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Experience ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Report created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report an experience",
        "tags": [
          "Reports"
        ]
      }
    },
    "/trips/{id}/report": {
      "post": {
        "operationId": "ReportsController_reportTrip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Trip ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Report created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report a trip",
        "tags": [
          "Reports"
        ]
      }
    },
    "/users/{id}/report": {
      "post": {
        "operationId": "ReportsController_reportUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Report created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report a user",
        "tags": [
          "Reports"
        ]
      }
    },
    "/admin/reports": {
      "get": {
        "operationId": "ReportsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Expérience",
                "Trajet",
                "Utilisateur",
                "Lieu",
                "Commentaire"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports returned."
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reports (min 10 reports per entity)",
        "tags": [
          "Reports"
        ]
      }
    },
    "/admin/reports/{id}/status": {
      "patch": {
        "operationId": "ReportsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update report status",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Swagger pour ton backend NestJS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BadRequestResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Bad Request"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "ForbiddenResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 403
          },
          "message": {
            "type": "object",
            "example": "Forbidden"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "enum": [
              "Expérience",
              "Trajet",
              "Utilisateur",
              "Lieu",
              "Commentaire"
            ],
            "example": "Expérience"
          },
          "entityId": {
            "type": "string",
            "example": "uuid-of-entity"
          },
          "motif": {
            "type": "string",
            "example": "Inappropriate content"
          }
        },
        "required": [
          "entityType",
          "entityId",
          "motif"
        ]
      },
      "UpdateReportStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "Nouveau",
              "Traité"
            ],
            "example": "Traité"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}